Bump/remove expiring Attribution Reporting metrics
We retain the following metrics, as they may still be useful during the
deprecation process:
- Conversions.AttributionReportDebugKeyUsage
- Conversions.OsRegistrationItemsPerBatch
- Conversions.Storage.Sql.InitStatus2
- Conversions.TimeFromTriggerToReportSentSuccessfully

OBSOLETE_HISTOGRAM[Conversions.AggregatableReport.SchedulerReportDelay]=We
have already removed the equivalent event-level metric.
OBSOLETE_HISTOGRAM[Conversions.AggregatableReport.NumContributionsPerReport2]=We
are no longer considering changes to the number of contributions.
OBSOLETE_HISTOGRAM[Conversions.OsRegistrationItemsPerBatch]=We are no
longer considering changes to batch size limits.
OBSOLETE_HISTOGRAM[Conversions.OsRegistrationsBufferFlushReason]=We are
not considering increasing the timeout here.
OBSOLETE_HISTOGRAM[Conversions.UniqueReportingOriginsPerSiteForAttribution]=We
are no longer considering changes to the limit here.
OBSOLETE_HISTOGRAM[Conversions.DebugReport.ReportRetrySucceedAggregatable]=We
are not considering changes to this behavior.
OBSOLETE_HISTOGRAM[Conversions.DebugReport.ReportRetrySucceedEventLevel]=We
are not considering changes to this behavior.
OBSOLETE_HISTOGRAM[Conversions.ReportRetrySucceedEventLevel]=We are not
considering changes to this behavior.
OBSOLETE_HISTOGRAM[Conversions.NoContextID.ReportRetrySucceedAggregatable2]=We
are not considering changes to this behavior.
OBSOLETE_HISTOGRAM[Conversions.ReportRetrySucceedAggregatable2]=We are
not considering changes to this behavior.
OBSOLETE_HISTOGRAM[Conversions.ContextID.ReportRetrySucceedAggregatable2]=We
are not considering changes to this behavior.

Bug: 451610074, 453662438, 459474223, 461500575
Change-Id: I0ef28eb3661b4b50d377f8500e048ba7a5cc780b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7169203
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547193}