Improve memory safety of cc::ScrollJankV4Frame::args
The cc::ScrollJankV4Frame::args field was a raw reference to
viz::BeginFrameArgs. The issue with that is that, if used incorrectly
(if the lifetime of the args was shorter than that of the frame), it
could lead to use-after-free errors.

On the other hand, storing viz::BeginFrameArgs by value is prohibitively
expensive (over 100 bytes). This CL modifies cc::ScrollJankV4Frame::args
to store a *subset* of viz::BeginFrameArgs' fields by value, i.e. we
keep only the fields which are relevant to the scroll jank v4 metric.

Bug: b:462067445
Change-Id: I3e7131d437b2375f22f0d714e85fdb1564875f79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7167092
Reviewed-by: Kartar Singh <kartarsingh@google.com>
Commit-Queue: Petr Cermak <petrcermak@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547196}