[a11y] Update TextInputState.getSurroundingTextInternal to include the offset.
The offset returned by `getSurroundingText` was previously hardcoded as
-1. This update includes a more accurate calculation of this value,
subtracting the length of the surrounding text to the left of the cursor
from the selection's starting index (the cursor's current index).

The corresponding test has also been updated to validate the offset.

Bug: 419335930
Change-Id: I21d2e0349f7fd536313d183851ca51766e614c40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7147322
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Carlos Marcelo Barrera Nolasco <carlosbarrera@google.com>
Cr-Commit-Position: refs/heads/main@{#1547205}