YourSavedInfo: Align category visibility with Autofill AI eligibility
When a user is not eligible for Autofill AI, but have some data already
saved, the 'Your Saved Info' page incorrectly displayed all Autofill
AI-related categories, such as 'Identity docs' and 'Travel',
even those for which the user had no data saved.

This was because `showAutofillAiControl` is set to true if user has some
data. It made other chips to be displayed as if the user was eligible.

This change corrects it by looking at `userEligibleForAutofillAi`,
which is disabled even when user has some data, allowing our logic
to calculate the visibility for every data chip individually, taking
into account number of saved items.

Bug: 461714590
Change-Id: I2fcd622338f3c692e1ac13cad0aea30a6a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7165025
Reviewed-by: Julia Sobiech <jsobiech@google.com>
Commit-Queue: Ireneusz Szulc <ireneuszszulc@google.com>
Cr-Commit-Position: refs/heads/main@{#1547254}