[iOS] Add navigation bar in tabs picker
This CL introduces a navigation bar to the tab picker screen in the
Composebox feature for iOS.

Previously, the tab picker had a custom header and a bottom toolbar for
actions. This change refactors the UI to use a UINavigationController,
which provides a standard navigation bar at the top of the screen.

The new navigation bar includes: A "Cancel" button on the left to
dismiss the tab picker. An "Add" button on the right to confirm the
selection of tabs.

See iOS 18 video:
https://drive.google.com/file/d/1dDVppIC4SuFwXiwglm0i3YnDE_hnKI7q/view?usp=drive_link

See iOS 26 video:
https://drive.google.com/file/d/18YhJzP4S1z0kCfARauF1IT2s-aJg9tZF/view?usp=drive_link

See iOS 26 scrolling with a lot of tabs:
https://drive.google.com/file/d/1W1LMxHIFsCvRaAAejyivNB5MvMBVMRjE/view?usp=sharing

Fixed: 458025976, 458026095
Change-Id: Id0d0242d79df26a8e6b14c1b84965aa11c81c642
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7131339
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547267}