Updating CorpMessagingPlayground to use messaging authz token
This CL is an incremental change to add 'username' and the concept of
a messaging authz token to the Corp messaging playground and related
classes. In order the keep the CLs small, I'm limiting the churn to that
and will send a follow-up for other changes (such as using the new burst
and ping-pong proto messages).

Bug: 359620500
Change-Id: I3f2ac4613f216bf05c1d0b1928c399bb7f0c2088
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7171177
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547276}