[TBHSP] Reparent tap strip on touch mode change.
With the removal of webui tab strip, touch/tablet mode on ChromeOS has a
few views painted overtop webcontents when the top container is hidden
during scroll.

This occurs because the tabstrip (non-webui) is parented to browser_view
instead of top_container, so the shown_ratio_ is not being applied to
the any layers in tab_strip_view. To fix this, we can reparent
tab_strip_view to top_container whenever the touch mode changes to
enabled.

Bug: 448166048
Change-Id: If738d4f42f0a154569d65d943f8bfb728e529df1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7164237
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Vince Lugli <lugli@google.com>
Cr-Commit-Position: refs/heads/main@{#1547290}