[lens] Create LensResultsPanelRouter to abstract side panel logic
This CL introduces the `LensResultsPanelRouter` class to serve as an
abstraction layer for interactions with the results panel. This allows
`LensOverlayController` to communicate with the results panel without
directly depending on the `LensOverlaySidePanelCoordinator`.

The router currently forwards calls to the side panel coordinator while
ensuring the search controller is in a valid state. In the future, it
will route these calls to the next panel

Key changes:
- Adds `LensResultsPanelRouter` class.
- Initializes the router within `LensSearchController`.
- Updates `LensOverlayController` to use the router for some operations.

Bug: 457447805
Change-Id: If62488aa212f69a2871b86cfc80a44b3f8fd160a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7163939
Commit-Queue: Juan Mojica <juanmojica@google.com>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Auto-Submit: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1547324}