Fix crash in GlicFloatingUi::MaybeSetWidgetCanResize
This change fixes a crash that occurs when the GlicFloatingUi is closed. The crash is caused by a call to MaybeSetWidgetCanResize while the browser is shutting down. At this point, the widget delegate has already been released.

The fix adds a null check at the beginning of MaybeSetWidgetCanResize to avoid the crash.

Bug: 461877512
Change-Id: Icf112c9f98b49556e2cf1e43eef641fface23d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172195
Commit-Queue: Dan Harrington <harringtond@chromium.org>
Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547340}