[fuchsia] Finish push messaging api implementation
This CL largely finishes the push messaging api, there are several
TODOs, but none of them impact the functionality.

The platform notification is not a must-have in fuchsia_web / webengine,
but it can be meaningful for debugging purpose. To try the e2e push
messaging api behavior locally, we can inspect the web-engine stdout to
retrieve the pushed messages from the PlatformNotificationServiceImpl.
E.g. I use https://simple-push-demo.vercel.app to try it, and the output
looks like,

```
[352115.929056][26760547][26760549][web_instances:web_instances_63273b01-b398-43b3-9aa1-d2d523eb3ef0][web_engine_exe] WARN: [fuchsia_web/webengine/browser/platform_notification_service_impl.cc(25)] DisplayNotification p#https://simple-push-demo.vercel.app/#00 from https://simple-push-demo.vercel.app/ in https://simple-push-demo.vercel.app/
[352115.929106][26760547][26760549][web_instances:web_instances_63273b01-b398-43b3-9aa1-d2d523eb3ef0][web_engine_exe] WARN: [fuchsia_web/webengine/browser/platform_notification_service_impl.cc(27)] Received Payload: Push data: 'blablabla'
```

I'd expect this change to increase the binary size, so this CL doesn't
enable the WEB_ENGINE_ENABLE_PUSH_MESSAGING_API. I will do it
separately.

Bug: 424479300
Change-Id: If7206ad7d9fe435722a288d3eac6df5f2bdab25f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7047169
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Zijie He <zijiehe@google.com>
Cr-Commit-Position: refs/heads/main@{#1547357}