Plumb through host_public_key and update OpenChannel struct
The HostOpenChannel struct is missing a couple of important fields
so this CL adds them and fills them in. I will need to follow up
with a src-internal CL to update the proto conversion code to use
them but it's easier to add them before having to roll deps.

Bug: 359620500
Change-Id: I0d25974600df948b9af213ab2b3ba9266f0f8ceb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7171972
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547381}