[iOS][CX] Implement resume in passkey importer
Implements the second part of the passkey importer API that:
* stores previously cached valid passkeys in the passkey model
* stores subset of previously cached conflicting passkeys, that were
chosen by the user in a conflict resolution UI (if there were any)

For the conflicting passkeys, API accepts vector of int ids, describing
the element of the conflicting passkeys vector that should be imported.
This id information will probably need to be passed to the caller in a
follow-up in the initial processing callback, so that the caller does
not need to retain the order on the conflict UI list.

Bug: 458337350
Change-Id: I857d580c06eb077ac131f71651d146e0f1702e6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7162393
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1547429}