[GlicMI] Make sure that the webui_contents has a compositor for screenshot
There is a mechanism in place to ensure that screeenshots come from the
instant that it's requested instead of at some point in the future. But
that only works if the webcontents has a compositor. We were removing
GlicView from the view hierarchy before requesting the screenshot, so
it wasn't getting that mechanism. This change adjusts to request the
screenshot before the GlicView is removed from the hierarchy.

Bug: b:461827011
Change-Id: If42f3f5aba792bc1eb1d58bc58a0c13b6a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172987
Reviewed-by: Basia Zimirska <basiaz@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547470}