[Contextual tasks] Update cache on task change
When the current side panel is changed to a new task or an existing task
Update the task_id of the current WebContents. If there is already
WebContents with the same task_id, remove it from the cache since we
only maintain 1 WebContents per task.

Bug: 461910573
Change-Id: I07d479d9c80eda2e0e469c3c14a2b4ae23a2e534
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7165629
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Yuheng Huang <yuhengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547495}