[WaaP] Improve accuracy of `InputTo<Action>` metrics
For the following metrics, this CL moves the logging of their start
mark from TypeScript into C++, i.e.
`InitialWebUIPageLoadMetricsObserver::OnUserInput()`. This is
necessary as (1) in TS there is significant delay between the timing
of input event and when the event callback is triggered. (2) The
TS event.timeStamp can't be converted back to base::TimeTicks.

The PageLoadMetricsObserver approach utilizes the fact that it
receives InputEvent from `MetricsWebContentsObserver`, which is a
`content::RenderWidgetHost::InputEventObserver` and guaranteed to
receive raw InputEvent data from `RenderWidgetHost`.

- `InitialWebUI.ReloadButton.InputToReload.MouseRelease`
- `InitialWebUI.ReloadButton.InputToStop.MouseRelease`

Bug: 448794588
Change-Id: I07d8e8d39c54868164a6296aeeeae1a1534ae0b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7167192
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Commit-Queue: Ming-Ying Chung <mych@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547585}