Enable SharedDictionaryCache by default with 2-entries
This enables the LRU memory cache for compression dictionaries
used by compression dictionary transport (dcb and dcz content encoding)
for dictionaries that are used for document requests (where dictionaries
are likely to be reused).

The experiment with 10% of stable for each of the three configs
(1, 2 and 5 cache entries) showed that all of the configurations had
a positive impact on performance with no negative side-effects at that
the 2-entry configuration performed the best.

Bug: 411711704
Change-Id: Ic33cd22abb7444aab3c04f0c01a9c02d8d4a9341
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7169887
Commit-Queue: Patrick Meenan <pmeenan@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547597}