WebNN: Add support for EP blocklist and command line override
This change adds a small feature for allowing the user to enable all IHV
Execution Providers (EPs) that are in WebNN's known EP list. We
introduce the concept of an enablement state to the EP data structure
and the scaffolding necessary to process the new command line option
that controls it.

A user would invoke this via the following command line, e.g.:
`--webnn-ort-ignore-ep-blocklist`

This will cause WebNN to treat all EPs in the list as enabled,
regardless of their enablement state. This can be useful for testing
purposes (e.g., running conformance tests and demos for an EP that is
being stabilized for WebNN scenarios).

I also updated the EP list to track some new EPs which are not enabled
by default (but via this switch, now they can be!).

How tested: I used a private build to try turning on the Vitis EP; I
could observe the EP module being loaded as part of attempting a demo.

Change-Id: I85a54eb84411fa46701724d15c5ec1626dbf68b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7133323
Reviewed-by: Hu, Ningxin <ningxin.hu@intel.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Commit-Queue: Hu, Ningxin <ningxin.hu@intel.com>
Cr-Commit-Position: refs/heads/main@{#1547608}