[Blink] Inline FlushCanvas() in CRPBitmap::DoExternalDrawAndSnapshot()
Part of the continued move to have this method be standalone as we look
to then trim it down and decouple it from CRPBitmap.

The code inlined is from here [1]. For ease of review, I did the minimal
changes when inlining it; I'll streamline it in followup.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_resource_provider.cc;l=1654?q=CanvasResourceProvider::FlushCanv&sq=&ss=chromium

Bug: 352263194
Change-Id: Idf6d26f1d338530eb62d7447d2a66d8a6ca2c810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172648
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547663}