[Blink] Eliminate test needing to configure TestSII specially for GMBs
Instead just have the test create the GMB handle the way that TestSII
does in the "configure specially" path in addition to extending TestSII
to strip out external sampling being set if appropriate (which already
happens within the "configure specially" path as well [1]).

This will pave the road to getting rid of the need for this special
configuration altogether.

[1] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/client/client_shared_image.cc;l=720-725?q=ClientSharedImage::CreateFo&ss=chromium

Bug: 404905709
Change-Id: I27ec1fe253f9436df70bc050d9af889e6effafd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172282
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547670}