[AutofillAi - M4] Create EntityDataManager.java to forward calls to CC
This method is similar to PersonalDataManager.java and will be used by
settings to communicate with EDM in the c++ side.

For now I added a simple method and all the boilerplate, methods to
retrieve, update and add entities will be added in a follow up.

Bug: 411324196
Change-Id: Ieeefcd776ae8612703d05836418f006465d7a9fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172417
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1547727}