[<geolocation> element] Break web-facing dependency on `<permission>`.
This CL causes `HTMLGeolocationElement` to use `HTMLElement` as its
prototype, as opposed to `HTMLPermissionElement`.

This CL doesn't aim to change any behavior at the C++ level, nor does it
unwind any of the dependency between runtime-enabled flags, but it's an
initial step towards making the `HTMLGeolocationElement` interface stand
apart from `HTMLPermissionElement` such that it will be possible to ship
the one without the other.

Bug: 461543463
Change-Id: I25939a66d3c34ba3701ea678148323ee0b47f0dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7167055
Reviewed-by: Thomas Nguyen <tungnh@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547729}